sweet alert

+++ 3-cart-app/src/app/components/cart-app.ts
@@ -5,6 +5,8 @@ import { Navbar } from './navbar/navbar';
 import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../services/sharing-data';
 
+import Swal from 'sweetalert2';
+
 @Component({
   selector: 'cart-app',
   imports: [RouterOutlet, Navbar],
@@ -39,6 +41,12 @@ export class CartAppComponent implements OnInit {
       this.router.navigate(['/cart'], {
         state: { items: this.items, total: this.total },
       });
+
+      Swal.fire({
+        title: 'Shopping cart',
+        text: 'Producto añadido',
+        icon: 'success',
+      });
     });
   }
 
@@ -50,13 +58,31 @@ export class CartAppComponent implements OnInit {
   }
   removeProduct() {
     this.sharingDataService.removeEventEmitter.subscribe((id) => {
-      this.items = this.service.removeProduct(id);
-      this.calculateTotal();
-
-      this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
-        this.router.navigate(['/cart'], {
-          state: { items: this.items, total: this.total },
-        });
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.items = this.service.removeProduct(id);
+          this.calculateTotal();
+          Swal.fire({
+            title: 'Deleted!',
+            text: 'Your file has been deleted.',
+            icon: 'success',
+          });
+          this.router
+            .navigateByUrl('/', { skipLocationChange: true })
+            .then(() => {
+              this.router.navigate(['/cart'], {
+                state: { items: this.items, total: this.total },
+              });
+            });
+        }
       });
     });
   }
